Convert emoji to ts
